+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-efcis
 
 0.2 (unreleased)
 ----------------
+- Add season as filter paremeter.
+
 - Add 'omschrijving' field to MeetStatus, serializer.
 
 - Added API for locations, including coloring of the points per value.